Create Kata  Home Report home for your next assignment TRAINING Practice Complete challenging Kata to earn honor and ranks. Re-train to hone technique Freestyle Sparring Take turns remixing and refactoring others code through Kumite CAREER Opportunities Find your next career challenge – powered by Qualified Jobs COMMUNITY Leaderboards Achieve honor and move up the global leaderboards Chat Join our Discord server and chat with your fellow code warriors Discussions View our Github Discussions board to discuss general Codewars topics ABOUT Docs Learn about all of the different aspects of Codewars Spencer06 Avatar 4 kyu 794 8 kyu Correct the mistakes of the character recognition software (8).py
Character recognition software is widely used to digitise printed texts. Thus the texts can be edited, searched and stored on a computer.

When documents (especially pretty old ones written with a typewriter), are digitised character recognition softwares often make mistakes.

Your task is correct the errors in the digitised text. You only have to handle the following mistakes:

S is misinterpreted as 5
O is misinterpreted as 0
I is misinterpreted as 1
The test cases contain numbers only by mistake.